Split large mapper files
