+++ internal/pkg/mapper/relations/local_save.go
@@ -20,13 +20,3 @@ func (m *relationsMapper) MapBeforeLocalSave(recipe *model.LocalSaveRecipe) erro
 	recipe.ChangedFields.Remove(`relations`)
 	return nil
 }
-
-// MapBeforeRemoteSave - modify changed fields.
-func (m *relationsMapper) MapBeforeRemoteSave(recipe *model.RemoteSaveRecipe) error {
-	if recipe.ChangedFields.Has(`relations`) {
-		// Relations are stored on the API side in config/row configuration
-		recipe.ChangedFields.Add(`configuration`)
-		recipe.ChangedFields.Remove(`relations`)
-	}
-	return nil
-}
