+++ internal/pkg/mapper/sharedcode/codes/local_load_test.go
@@ -10,23 +10,6 @@ import (
 	"github.com/keboola/keboola-as-code/internal/pkg/model"
 )
 
-func TestSharedCodeLegacyRemoteCodeContent(t *testing.T) {
-	t.Parallel()
-	context, row, _ := createTestFixtures(t, `keboola.snowflake-transformation`)
-	row.Content.Set(model.SharedCodeContentKey, "SELECT 1; \n  SELECT 2; \n ")
-
-	event := model.OnObjectsLoadEvent{
-		StateType:  model.StateTypeRemote,
-		NewObjects: []model.Object{row},
-		AllObjects: context.State.RemoteObjects(),
-	}
-	assert.NoError(t, NewMapper(context).OnObjectsLoad(event))
-
-	v, found := row.Content.Get(model.SharedCodeContentKey)
-	assert.True(t, found)
-	assert.Equal(t, []interface{}{"SELECT 1;", "SELECT 2;"}, v)
-}
-
 func TestSharedCodeLoadMissingFile(t *testing.T) {
 	t.Parallel()
 	targetComponentId := `keboola.python-transformation-v2`
